敵出現データのtimeを正数値型から浮動小数型へ変更

+++ unity_2d_stg/Assets/Scripts/Stg/Game/Enemy/StgEnemyLoadResourceStageJson.cs
@@ -29,7 +29,7 @@ public class StgEnemyLoadResourceStageJson {
                             launchData.enemy_type = (string)val;
                             break;
                         case "time":
-                            launchData.time = int.Parse(val);
+                            launchData.time = float.Parse(val);
                             break;
                         case "x":
                             launchData.x = float.Parse(val);
